refactor: fix problems
