+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from './types';
 
+export * from './client';
 export * from './result';
-export * from './querybuilder/query';
+export * from './query';
 export * from './client-repo';
-export * from './doofinder';
